@@ -138,12 +138,13 @@ class MergeRequests extends AbstractApi
     /**
      * @param int   $project_id
      * @param int   $mr_id
+     * @param array $parameters
      *
      * @return mixed
      */
-    public function showNotes($project_id, $mr_id)
+    public function showNotes($project_id, $mr_id, array $parameters = [])
     {
-        return $this->get($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($mr_id).'/notes'));
+        return $this->get($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($mr_id).'/notes'), $parameters);
     }
 
     /**
